notebooks/requirements.txt
jupyter-book
pyyaml
